가독성 있는 함수명으로 수정 및 스타일시트 세분화
