+++ src/styles/SignUp/SignUp.styled.js
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+const Container = styled.div`
+  width: 500px;
+  margin: 100px auto;
+`;
+const Title = styled.h1`
+  font-size: 20px;
+  font-weight: bold;
+  margin-bottom: 20px;
+  text-align: center;
+`;
+
+export { Container, Title };
